move editor into lib file, fix ChatArea
